+++ tests/unit/events/test_event_listener.py
@@ -1,12 +1,12 @@
 from unittest.mock import Mock
 import pytest
 from griptape.structures import Pipeline
-from griptape.tasks import ToolkitTask, ActionSubtask
+from griptape.tasks import ToolkitTask, ActionsSubtask
 from griptape.events import (
     StartTaskEvent,
     FinishTaskEvent,
-    StartActionSubtaskEvent,
-    FinishActionSubtaskEvent,
+    StartActionsSubtaskEvent,
+    FinishActionsSubtaskEvent,
     StartPromptEvent,
     FinishPromptEvent,
     StartStructureRunEvent,
@@ -26,7 +26,7 @@ class TestEventListener:
         pipeline = Pipeline(prompt_driver=MockPromptDriver(stream=True))
         pipeline.add_task(task)
 
-        task.add_subtask(ActionSubtask("foo"))
+        task.add_subtask(ActionsSubtask("foo"))
         return pipeline
 
     def test_untyped_listeners(self, pipeline):
@@ -58,8 +58,8 @@ class TestEventListener:
             EventListener(finish_prompt_event_handler, event_types=[FinishPromptEvent]),
             EventListener(start_task_event_handler, event_types=[StartTaskEvent]),
             EventListener(finish_task_event_handler, event_types=[FinishTaskEvent]),
-            EventListener(start_subtask_event_handler, event_types=[StartActionSubtaskEvent]),
-            EventListener(finish_subtask_event_handler, event_types=[FinishActionSubtaskEvent]),
+            EventListener(start_subtask_event_handler, event_types=[StartActionsSubtaskEvent]),
+            EventListener(finish_subtask_event_handler, event_types=[FinishActionsSubtaskEvent]),
             EventListener(start_structure_run_event_handler, event_types=[StartStructureRunEvent]),
             EventListener(finish_structure_run_event_handler, event_types=[FinishStructureRunEvent]),
             EventListener(completion_chunk_handler, event_types=[CompletionChunkEvent]),